Refactors RGBU export code to separate exporter model, reuses for single axis RGBU plot
