+++ client/src/app/modules/widget/components/widget-export-dialog/widget-export-model.ts
@@ -19,6 +19,9 @@ export type WidgetExportOption = {
   description: string;
   selected: boolean;
 
+  disabled?: boolean;
+  disabledText?: string;
+
   // Only for multiswitch
   options?: string[];
   selectedOption?: string;
